\name{wpfSA24204165-package}
\alias{wpfSA24204165-package}
\alias{wpfSA24204165}
\docType{package}
\title{FinalProject}
\description{\packageDescription{wpfSA24204165}}
\section{Package Content}{\packageIndices{wpfSA24204165}}
\author{
WuPengfeiSA24204165
Maintainer: WuPengfei <wpfpg68@mail.ustc.edu.cn>
}
\section{Maintainer}{\packageMaintainer{wpfSA24204165}}
\keyword{package}
